+++ rules/layout-formatting.cjs
@@ -1,6 +1,0 @@
-// https://eslint.org/docs/latest/rules/#layout--formatting
-
-/** @type {import("eslint").Linter.Config["rules"]} */
-module.exports = {
-  "line-comment-position": "error"
-};