+++ src/rules/layout-formatting.ts
@@ -1,0 +1,6 @@
+import type { Linter } from "eslint";
+
+// https://eslint.org/docs/latest/rules/#layout--formatting
+export = {
+  "line-comment-position": "error"
+} satisfies Linter.Config["rules"];